move mapping to separated file
